+++ src/layouts/Pane/styles/PaneStyles.js
@@ -1,28 +1,16 @@
 import styled from 'styled-components';
 import get from 'extensions/themeGet';
+import PaneRow from '../PaneRow';
 
 export default styled.article`
   display: flex;
   flex-direction: column;
   flex: 1 1;
-  padding: ${get('spacing.large')};
+  padding: 0 ${get('spacing.large')};
 
-  & > h1,
-  & > h2,
-  & > h3,
-  & > h4,
-  & > h5 {
-    margin: ${get('spacing.medium')} 0;
-  }
+  /* for panes within rows that are not on the top, add top padding */
 
-  & > section {
-    margin: ${get('spacing.large')} 0 0 0;
-
-    &:first-of-type {
-      margin-top: 0;
-    }
-    &:last-of-type {
-      margin-bottom: ${get('spacing.large')};
-    }
+  ${PaneRow}:not(:first-of-type) > & {
+    padding-top: ${get('spacing.large')};
   }
 `;